+++ harness/drift-reference.sh
@@ -28,6 +28,8 @@ twin() {
     JsonConfigSource.php)     echo "src/Composer/Config/JsonConfigSource.php" ;;
     RemoveCommand.php)        echo "src/Composer/Command/RemoveCommand.php" ;;
     VersionSelector.php)      echo "src/Composer/Package/Version/VersionSelector.php" ;;
+    RequireCommand.php)       echo "src/Composer/Command/RequireCommand.php" ;;
+    PackageDiscoveryTrait.php) echo "src/Composer/Command/PackageDiscoveryTrait.php" ;;
     Locker.php)               echo "src/Composer/Package/Locker.php" ;;
     PackageSorter.php)        echo "src/Composer/Util/PackageSorter.php" ;;
     RootPackageLoader.php)    echo "src/Composer/Package/Loader/RootPackageLoader.php" ;;